requirements.txt
fastapi==0.109.1
uvicorn==0.23.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.8
pydantic==2.4.0
pydantic-settings==2.2.0
alembic==1.12.0
python-dotenv==1.0.0
pytest
httpx
pytest-mock